Moved reset_db in run.py for now, and added to the context processor
